Param caching made optional, disabled by default
In some cases, the autopilot might not broadcast PARAM_VALUE
messages following a PARAM_SET from a controller. This means that
if a separate GCS instance changes a param, this software
won't find out without an explicit re-fetch. So, caching is disabled
by default and the module will re-fetch the param each time it
is needed. If caching is desired, it can be enabled in the param
module's init() function by passing True to use_cache in the class
constructor.