Use relative altitude in gps_odom topic
Rationale: the autopilot accepts most commands using relative altitude
(AGL wrt takeoff point). To provide external applications with absolute
altitude, the 'gps' topic still publishes GPS MSL. It is assumed that the
Odometry version of the publisher is more likely to be used for onboard
processing and therefore AGL is more useful.