Added guards to system time handling
There were cases where exceptions would get thrown because a negative
(invalid) time was being calculated. For instance, if the GPS time
hack was disabled and the aircraft was booted indoors, where a valid
timestamp was never generated on the autopilot. Fixed up the logic
to catch these cases and simply not perform the offending calculation.