fpr: initial commit
Added new module for Fence, Params, and Rally points (FPR). Uses
common underlying functions to get, set, and verify data items.
It is a work in progess; param get/set works, but fence and rally
points need additional implementation (code is pass-through or
commented out as appropriate).

Note that this module and the 'param' module should not be used
together, as they implement overlapping service names.