acs: new version service, calpress improvement
Added autopilot/version as a service that returns a bunch of info
about the autopilot capabilities, software versions, etc. Right
now, only the field .flight_custom is used - it contains the first
few characters of the Git hash for the current ArduPlane code.

Also improved the calpress retry mechanism a bit, based on how
the version retry works (can resend if it hasn't started yet).