core: allow re-init and better clock mgmt
Moderately extensive changes to how clocks are set and the AP reported
boot-time is managed. Logic for syncing clocks, setting time variables,
and setting up the MAVLink stream, has been shifted and consolidated
into ._init_stream() and ._init_time(sync=False). These can be used
to reinitialize a stream and timing once the bridge can been started.

STATUSTEXT messages are now examined in ._handle_statustext(msg),
and if the text 'Ready to FLY.' is seen, ._init_time() is called to
reset clock variables. This allows us to detect when the autopilot
reboots and to handle the discontinuity in its SYSTEM_TIME messages.
Note that SITL autopilots do not currently generate that STATUSTEXT
message, so this detection will not work in simulation.

In the process, also cleaned up the computation for syncing the local
clock, and eliminated the now-deprecated --track-time-delta option.