core: attempt reconnect on failure
If a serial port fails due to disappearing, the bridge will catch
it and attempt to rediscover and reconnect. If using auto-detect,
it will catch cases where the serial device name changes (e.g.,
/dev/ttyUSB0 becomes /dev/ttyUSB1 because USB-TTL adapter is
jostled and Linux doesn't let go of the old device before assigning
a new one).

Note, the reworking got rid of some of the baudrate selection logic,
so a non-default baudrate should always be specified. Future work
might be to set defaults for certain kinds of devices (ACM vs USB).