acs: Fixed surface demo override issues
It turns out that 0, not -1, is the magic number for undoing an
override on an RC channel. Combined that fix with the existing
trimming logic to account for the absence of an RC transmitter,
plus some minor tweaks to sleeps, and demos now appear to work
without permanently disabling manual control.